Initial commit
Create an array of M random integers. Create n processes and each of them makes a copy of the array, from a shared memory of size 1 integer. At the same time, each of the n processes calculates a moving average for the time it takes from writing the integer (each of M integers) in the shared memory, until the reader gets the integer.